@@ -1,5 +1,3 @@
-module github.com/rodkranz/fetch
+module github.com/jad21/fetch
 
-go 1.13
-
-require golang.org/x/net v0.0.0-20180801234040-f4c29de78a2a
+go 1.19
