@@ -3,18 +3,16 @@ package main
 import (
 	"log"
 
-	"github.com/rodkranz/fetch"
+	"github.com/jad21/fetch"
 )
 
-const url = "https://api.github.com/users/rodkranz"
+const url = "https://api.github.com/users/jad21"
 
 func main() {
-	f := fetch.NewDefault()
-	rsp, err := f.Get(url, nil)
+	rsp, err := fetch.Get(url, nil)
 	if err != nil {
 		log.Fatalf("could not fetch [%s] because: %s", url, err)
 	}
-
 
 	body, err := rsp.ToString()
 	if err != nil {
